Python-Backend/sentiment-analysis/pyproject.toml
[tool.poetry]
name = "sentiment-analysis"
version = "0.1.0"
description = ""
authors = ["Karl Duignan <karl.duignan@fabfitfun.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
fastapi = "^0.95.2"
transformers = "^4.29.2"
torch = "^2.0.1"
tensorflow = "^2.12.0"
flax = "^0.6.10"
SQLAlchemy = "^2.0.13"
python-dotenv = "^1.0.0"
aiomysql = "^0.1.1"
vaderSentiment = "^3.3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
